requirements.txt
anyio==4.7.0
certifi==2024.12.14
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
python-telegram-bot==21.10
pytubefix==8.3.0
sniffio==1.3.1
typing_extensions==4.12.2
validators==0.34.0
yt-dlp==2024.12.23
